pyproject.toml
[project]
name = "lawyergpt"
version = "0.1.0"
description = "Юридическая RAG-система на базе российского законодательства"
readme = "README.md"
requires-python = ">=3.10,<3.13"
authors = [{ name = "LawyerGPT" }]

dependencies = [
    "accelerate==1.7.0",
    "bert-score==0.3.13",
    "huggingface-hub==0.32.0",
    "langchain==0.3.25",
    "langchain-huggingface==0.2.0",
    "langchain-openai==0.3.18",
    "langchain-text-splitters==0.3.8",
    "matplotlib==3.10.3",
    "nltk==3.9.1",
    "numpy==2.2.6",
    "openai==1.82.0",
    "pandas==2.2.3",
    "pydantic==2.11.5",
    "psycopg2-binary==2.9.10",
    "qdrant-client==1.14.2",
    "sentence-transformers==4.1.0",
    "SQLAlchemy==2.0.41",
    "streamlit==1.45.1",
    "tiktoken==0.9.0",
    "tokenizers==0.21.1",
    "torch==2.6.0",
    "transformers==4.52.3",
    "python-dotenv>=1.0.1",
    "PyYAML>=6.0.1",
    "tqdm>=4.66.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "database*"]
